split property formatting into multiple files
